aerospace: Force Keepass window to be floating
Whenever Keepass perofrms an autotype, it minimizes the window, which
leads to other open windows being resized to take up the space the
Keepass app left. This can cause issues when browsers change size and
the layout of the website with the form being filled out also changes.

Setting Keepass windows to be floating prevents this issue.